Fix bug: DINO Amp

+++ mmdet/models/layers/transformer/dino_layers.py
@@ -335,7 +335,7 @@ class CdnQueryGenerator(BaseModule):
         chosen_indice = torch.nonzero(p < (self.label_noise_scale * 0.5)).view(
             -1)  # Note `* 0.5`
         new_labels = torch.randint_like(chosen_indice, 0, self.num_classes)
-        noisy_labels_expand = gt_labels_expand.scatter(0, chosen_indice,
+        noisy_labels_expand = gt_labels_expand.to(noisy_labels_expand.dtype).scatter(0, chosen_indice,
                                                        new_labels)
         dn_label_query = self.label_embedding(noisy_labels_expand)
         return dn_label_query
